tree: correct language name
At [1], we can see that the 'udl' variant simply uses
`tree_sitter_objectscript` as the exported symbol. If we call the
language `objectscript_udl` in `treesit-language-source-alist`, Emacs
will look for `tree_sitter_objectscript_udl`, which does not exist.

Rename the language to `objectscript` so that the correct symbol is
identified and found.

[1]: https://github.com/intersystems/tree-sitter-objectscript/blob/4d34cb2d4a4924f0b9c9524ffaba1b9c0644067e/setup.py#L12